Move CountingHandler in califonrium.core.test
